+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-   RadioButton: add `inputProps` prop to forward to the native input element.
+
+### Fixed
+
+-   RadioButton: fix helper text accessibility
+-   Switch: fix helper text accessibility
+
 ## [3.0.4][] - 2022-11-07
 
 ### Added